requirements.txt
Flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.7  # Add PostgreSQL adapter
python-dotenv==1.0.0
requests==2.31.0
ipfshttpclient==0.8.0a2
PyPDF2==3.0.1
werkzeug==2.3.7
groq==0.4.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
gunicorn==21.2.0

# Optional: For alternative IPFS services
# web3-storage==0.3.0
# pinata-python==0.1.0
